+++ packages/core/src/authFlows/__tests__/adfs.unit.spec.ts
@@ -21,6 +21,7 @@ describe('ADFS', () => {
   beforeEach(() => {
     window.history.pushState({}, '', '');
     nock.cleanAll();
+    sessionStorage.clear();
   });
 
   describe('login', () => {
@@ -114,15 +115,16 @@ describe('ADFS', () => {
         .mockResolvedValueOnce({
           accessToken,
           idToken,
-          expiredIn: Date.now() + 3600 * 1000,
+          expiresIn: Date.now() + 3600 * 1000,
         })
         .mockResolvedValueOnce({
           accessToken: updatedAccessToken,
           idToken: updatedIdToken,
-          expiredIn: Date.now() + 3600 * 1000,
+          expiresIn: Date.now() + 3600 * 1000,
         });
 
       const cdfTokenAfterLogin = await adfsClient.login();
+      sessionStorage.clear(); // force silent login
       const updatedCdfToken = await adfsClient.getCDFToken();
 
       expect(silentLogin).toHaveBeenCalledTimes(2);
@@ -136,16 +138,17 @@ describe('ADFS', () => {
         .mockResolvedValueOnce({
           accessToken,
           idToken,
-          expiredIn: Date.now() + 3600 * 1000,
+          expiresIn: Date.now() + 3600 * 1000,
         })
         .mockRejectedValueOnce(
           'Failed to acquire token silently due X-Frame-Options header deny'
         );
 
       const cdfTokenAfterLogin = await adfsClient.login();
-      const updatedCdfToken = await adfsClient.getCDFToken();
+      const anotherAdfsClient = new ADFS({ authority, requestParams });
+      const updatedCdfToken = await anotherAdfsClient.getCDFToken();
 
-      expect(silentLogin).toHaveBeenCalledTimes(2);
+      expect(silentLogin).toHaveBeenCalledTimes(1);
       expect(cdfTokenAfterLogin).toEqual(accessToken);
       expect(updatedCdfToken).toEqual(accessToken);
     });