+++ packages/stable/src/__tests__/api/visionApi.int.spec.ts
@@ -1,22 +1,30 @@
 // Copyright 2020 Cognite AS
 
-import { VisionExtractPostResponse } from '@cognite/sdk-playground';
-import CogniteClientPlayground from '../../cogniteClientPlayground';
-import { setupLoggedInClient as setupLoggedinPlaygroundClient } from '../testUtils';
+import { VisionExtractPostResponse } from '@cognite/sdk';
+import CogniteClient from '../../cogniteClient';
+import { BETA_FEATURES } from '../../api/vision/visionApi';
+import { setupLoggedInClient } from '../testUtils';
 
 describe('Vision API', () => {
   const TEST_IMAGE_ID = 4745168244986665;
-  let playgroundClient: CogniteClientPlayground;
+  let client: CogniteClient;
+  let consoleSpy: jest.SpyInstance;
   let extractJob: VisionExtractPostResponse;
+  let extractBetaJob: VisionExtractPostResponse;
 
   beforeAll(async () => {
     jest.setTimeout(2 * 60 * 1000); // timeout after 2 minutes
-    playgroundClient = setupLoggedinPlaygroundClient();
+    consoleSpy = jest.spyOn(console, 'warn').mockImplementation();
+    client = setupLoggedInClient();
 
-    extractJob = await playgroundClient.vision.extract(
+    extractJob = await client.vision.extract(
       ['TextDetection'],
-      [{ fileId: TEST_IMAGE_ID }]
+      [{ fileId: TEST_IMAGE_ID }],
+      { textDetectionParameters: { threshold: 0.4 } }
     );
+    extractBetaJob = await client.vision.extract(BETA_FEATURES, [
+      { fileId: TEST_IMAGE_ID },
+    ]);
   });
 
   test('extract', async () => {
@@ -28,14 +36,34 @@ describe('Vision API', () => {
     expect(extractJob.items).toEqual([
       { fileId: TEST_IMAGE_ID, fileExternalId: 'vision_extract_test_image' },
     ]);
+    expect(extractJob.parameters).toBeDefined();
+    expect(extractJob.parameters!.textDetectionParameters).toEqual({
+      threshold: 0.4,
+    });
+  });
+
+  test('extract using beta feature', async () => {
+    // Check that the beta flag is correctly used.
+    // Unfortunately, 'cdf-version' is not listed in
+    // access-control-allow-headers, which means we cannot explicitly check for
+    // that the 'cdf-version' entry is set to 'beta'. However, what we instead
+    // can do is to check that the API does not return 400 when a beta feature
+    // is sent.
+    const metadata = client.getMetadata(extractBetaJob);
+    expect(metadata).toBeDefined();
+    expect(metadata?.status).toEqual(200);
+    expect(consoleSpy).toBeCalledWith(
+      `Features '${BETA_FEATURES}' are in beta and are still in development`
+    );
+    // Only care that the job is queued with the correct feature. The other
+    // properties are checked for in the test above.
+    expect(extractBetaJob.status).toEqual('Queued');
+    expect(extractBetaJob.features).toEqual(BETA_FEATURES);
   });
 
   describe('retrieve extract job', () => {
     test('waitForCompletion=false', async () => {
-      const result = await playgroundClient.vision.getExtractJob(
-        extractJob.jobId,
-        false
-      );
+      const result = await client.vision.getExtractJob(extractJob.jobId, false);
       expect(result.status == 'Queued' || result.status == 'Running').toBe(
         true
       );
@@ -45,16 +73,13 @@ describe('Vision API', () => {
     });
     test('waitForCompletion=true, should timeout', async () => {
       await expect(
-        playgroundClient.vision.getExtractJob(extractJob.jobId, true, 1000, 0)
+        client.vision.getExtractJob(extractJob.jobId, true, 1000, 0)
       ).rejects.toThrowError(
         `Timed out while waiting for vision job to complete.`
       );
     });
     test('waitForCompletion=true', async () => {
-      const result = await playgroundClient.vision.getExtractJob(
-        extractJob.jobId,
-        true
-      );
+      const result = await client.vision.getExtractJob(extractJob.jobId, true);
       expect(result.status).toEqual('Completed');
       expect(result.jobId).toEqual(extractJob.jobId);
       expect(result.createdTime).toEqual(extractJob.createdTime);
@@ -65,7 +90,9 @@ describe('Vision API', () => {
       // we care in the following checks is that the *data structure* is
       // correctly filled.
       expect(result.parameters).toBeDefined();
-      expect(result.parameters!.textDetectionParameters).toBeDefined();
+      expect(result.parameters!.textDetectionParameters).toEqual({
+        threshold: 0.4,
+      });
 
       expect(result.items?.length).toBeGreaterThan(0);
       const resultItem = result.items![0];